feat: rename to `smock-foundry`
